refactor(backend): extract cost-summary aggregator into its own module + fix Avg/run denominator
Two review threads addressed in one commit (they touch the same models):

1. ntindle flagged that the cost-summary models and aggregator don't belong
   in `data/execution.py`. Move `UserAgentCostRollup`, `UserTopRun`,
   `UserDailyCost`, `UserExecutionCostSummary`, and `get_user_cost_summary`
   into a dedicated `backend/data/execution_cost_summary.py` (matches the
   existing `execution_cost_summary_test.py` test file name). Update
   imports in `api/features/v1.py` and both test files; drop the now-unused
   `asyncio` and `date` imports from `execution.py`.

2. Sentry flagged that the frontend "Avg / run" tile divides total_cents by
   run_count — which includes zero-cost runs in the denominator, deflating
   the displayed average. Add a `billable_run_count` column to the totals
   query (`COUNT(*) FILTER (WHERE cost > 0)`), surface it on the response,
   and divide by it in `CostsBreakdown.tsx`. "Runs" tile keeps the existing
   semantics so the count of all executions stays visible.

Tests:
- Extend `test_aggregates_by_agent_and_top_runs` with a zero-cost run
  asserting `run_count=6, billable_run_count=5`
- Add `test_empty_window` assertion on the new field
- New frontend test "computes Avg / run from billable_run_count, not run_count"
- Surgical patch to `openapi.json` (new required field), generated msw mocks
  updated locally (gitignored — CI regenerates from the patched spec)
