+++ autogpt_platform/frontend/src/app/(platform)/library/__tests__/briefing.test.tsx
@@ -73,6 +73,7 @@ function setupHandlers({ totalCents }: { totalCents: number }) {
     getGetV1UserCostSummaryMockHandler({
       total_cents: totalCents,
       run_count: totalCents > 0 ? 3 : 0,
+      billable_run_count: totalCents > 0 ? 3 : 0,
       failed_cost_cents: 0,
       by_agent:
         totalCents > 0
